Change indexes into tile arrays to be stored as bytes to save memory.
The index value needs only be big enough to handle all defined terrain types. This is a low number so we can save memory by defining it as a byte. This particularly saves memory for the CustomTerrain field in the Map class, which defines a cell layer for the map using tile indexes, so we can reduce the size of that layer 4x as a result.

+++ OpenRA.TilesetBuilder/FormBuilder.cs
@@ -385,7 +385,7 @@ namespace OpenRA.TilesetBuilder
 			ushort cur = 0;
 			foreach (var tp in surface1.Templates)
 			{
-				var tiles = new int[tp.Width * tp.Height];
+				var tiles = new byte[tp.Width * tp.Height];
 				foreach (var t in tp.Cells)
 				{
 					var ttype = TerrainType[surface1.TerrainTypes[t.Key.X, t.Key.Y]].Type;
